# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "e-Demic API Documentation",
        "description": "API endpoints and descriptions for e-Demic Laravel API",
        "contact": {
            "name": "Developer : Sreejith B S",
            "email": "sreejith.bs@quinoid.in"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/messages/{index}": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "List all messages",
                "description": "API to list all messages, such as alerts, announcements etc on latest first basis. Default limit set as **10** messages per page. <br><br>Accepts `index` for lazy load messages.",
                "operationId": "messages",
                "parameters": [
                    {
                        "name": "index",
                        "in": "path",
                        "description": "Page index to fetch messages",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/user/updateDeviceToken": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Update device token for push notification",
                "description": "API to update device token for push notification. <br><br>Accepts either `android_device_token` or `ios_device_token`.",
                "operationId": "updateDeviceToken",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "android_device_token": {
                                        "type": "string"
                                    },
                                    "ios_device_token": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "android_device_token": "654C4DB3-3F68-4969-8ED2-80EA16B46EB0",
                                    "ios_device_token": ""
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/doctor/fetchProfile/{doctor_id}": {
            "get": {
                "tags": [
                    "Doctor Users"
                ],
                "summary": "Fetch doctor user profile details",
                "description": "API to fetch doctor user profile details using the ID from doctor's Profile QR Code. <br><br>Accepts `doctor_id`.",
                "operationId": "fetchDoctorProfile",
                "parameters": [
                    {
                        "name": "doctor_id",
                        "in": "path",
                        "description": "Doctor ID to fetch profile",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/diseases": {
            "get": {
                "tags": [
                    "Diseases"
                ],
                "summary": "List all diseases",
                "description": "API to list all diseases.",
                "operationId": "diseases",
                "responses": {
                    "200": {
                        "description": "Successful Operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/user/diseaseDiagnosed": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Submit disease diagnosed patient details",
                "description": "API to submit disease diagnosed patient details to be saved in database.",
                "operationId": "diseaseDiagnosed",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "disease_code": {
                                        "type": "integer",
                                        "example": 6001
                                    },
                                    "stage_code": {
                                        "type": "integer",
                                        "example": 5001
                                    },
                                    "diagnosed_date_time": {
                                        "type": "string",
                                        "example": "2020-04-08 07:23:37"
                                    },
                                    "location_logs": {
                                        "type": "array",
                                        "items": {
                                            "example": {
                                                "date_time": "2020-04-08 07:23:37",
                                                "latitude": "51.528308",
                                                "longitude": "-0.131847",
                                                "address": "8-14 Eversholt St, Kings Cross, London NW1 1DG, UK"
                                            }
                                        }
                                    },
                                    "suspected_users_id": {
                                        "type": "object",
                                        "example": [
                                            1,
                                            2
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/patients": {
            "get": {
                "tags": [
                    "Map"
                ],
                "summary": "List all patient details for map plotting",
                "description": "API to list all patient details for map plotting.",
                "operationId": "diseases",
                "responses": {
                    "200": {
                        "description": "Successful Operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}